+++ CHANGELOG.md
@@ -31,6 +31,9 @@
   to specify that the rule should only be applied to `if` statements.  
   [Cihat Gündüz](https://github.com/Dschee)
   [#2307](https://github.com/realm/SwiftLint/issues/2307)
+* Add support for globs in `excluded` file paths  
+  [Keith Smiley](https://github.com/keith)
+  [#2316](https://github.com/realm/SwiftLint/pull/2316)
 
 * Add `private_only` configuration to `prefixed_toplevel_constant` rule  
   [Keith Smiley](https://github.com/keith)